Refactoring

+++ include/ICMPController.hpp
@@ -29,7 +29,7 @@ public:
     }
 
     void echo_request(const IPAddress & addr, uint16_t id, uint16_t ttl);
-    std::tuple<std::set<IPAddress>, int> echo_reply(uint16_t id, uint16_t ttl);
+    std::tuple<std::set<IPAddress>, ssize_t> echo_reply(uint16_t id, uint16_t ttl);
 
 private:
     IPAddress recv_echo(uint16_t id, uint16_t ttl);
